// src/types/db.d.ts
export interface User {
    name: string
    email: string
    image: string
    id: string
}

interface Chat {
    id: string
    message: Message[]
}

export interface Message {
    id: string
    senderId: string
    receiverId: string
    text: string
    timestamp: number
}

interface FriendRequest {
    id: string
    senderId: string
    receiverId: string
}